Build without clean, clean when args provided

+++ build.sh
@@ -1,12 +1,11 @@
 sh setup.sh
-rm -rf build
 
-mkdir -p build
+[ $1 = "--clean" ] && rm -rf build
+[ ! -d "build" ] && mkdir -p build
+
 cd build
 cmake ..
 make -j6
 cd ..
 
 mv build/predict .
-
-rm -rf build
